backend_patch/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
pandas==2.2.2
pandas_ta==0.3.14b0
python-dotenv==1.0.1
cachetools==5.3.3
